pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "security-cam"
version = "0.1.0"
description = "IP Camera RTSP Live Viewer"
authors = [{name = "Daniel Brutti", email = "danielbrutti@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opencv-python",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "pycln",
    "pre-commit",
]

[tool.black]
line-length = 79
target-version = ['py38']

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203"]

[tool.pycln]
all = true 

[tool.hatch.build.targets.wheel]
packages = ["src/security_cam"] 